+++ src/qcoder/__init__.py
@@ -1,3 +1,3 @@
 __all__ = []
-__version__ = "0.6.0a11"
+__version__ = "0.6.0a12"
 file = __file__